requirements.txt
streamlit
langchain-community
torch
pandas
numpy
openai
transformers
wikipedia
numexpr